+++ libraries/ghc-experimental/src/GHC/Stack/Annotation/Experimental.hs
@@ -82,9 +82,6 @@ import GHC.Internal.Stack.Annotation
 -- the pure variations can behave in ways that are hard to predict.
 --
 -- See Note [Stack annotations in pure code] for more details.
---
--- At last, stack annotations are tricky to use with 'error'.
--- See Note [Pushing annotation frames on 'error'] for why this is the case.
 
 -- Note [Stack annotations in pure code]
 -- ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -96,57 +93,50 @@ import GHC.Internal.Stack.Annotation
 -- For example:
 --
 -- @
---  annotateStackShow (5 @Int) (fib 20 + throw (ErrorCall "Oh no!"))
+--  annotateStackShow (5 @Int) (fib 20 + error "Oh no!")
 -- @
 --
--- Without forcing the result of @(fib 20 + throw (ErrorCall "Oh no!"))@, the computation
+-- Without forcing the result of @(fib 20 + error "Oh no!")@, the computation
 -- will simply return a thunk, and the stack annotation would be popped off the stack.
 -- Once the thunk is evaluated, the exception is raised, but no stack annotation will be found!
--- If we force the result of @(fib 20 + throw (ErrorCall "Oh no!"))@, then the stack
+-- If we force the result of @(fib 20 + error "Oh no!")@, then the stack
 -- annotations remain on the stack, and are displayed in the stack trace.
 --
 -- Naturally, this only holds if no imprecise exceptions are thrown during evaluation of any
 -- nested value, for example in 'annotateStackShow 5 (Just $ throw (ErrorCall "Oh no!"))', the
 -- stack trace will not include the value @5@.
 --
--- See how we preferred @throw (ErrorCall ...)@ over @error@?
--- See Note [Pushing annotation frames on 'error'] for why we do this.
-
--- Note [Pushing annotation frames on 'error']
--- ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
--- Examples so far have not been using 'error' at all.
--- The reason is that 'error' is extraordinarily difficult to use correctly with stack annotation frames.
--- See Note [Capturing the backtrace in throw] for a detailed discussion of how 'throw'
--- manages to capture 'Backtraces'.
---
--- Long story short, 'error' does not do the same thing as 'throw' and is subtly different
--- in terms of evaluation, cause it to bypass the stack annotation frames, especially in
--- pure code.
---
--- However, even in 'IO' code, it is difficult to use 'error' and obtain stack annotation frames
--- close to the call site due to the same issue of laziness and backtrace collection.
---
--- This means, right now, if you want to reliably capture stack frame annotations,
--- in both pure and impure code, prefer 'throw' and 'throwIO' variants over 'error'.
 
 -- ----------------------------------------------------------------------------
 -- Annotations
 -- ----------------------------------------------------------------------------
 
+
+-- | A 'String' only annotation with an optional source location.
 data StringAnnotation where
-  StringAnnotation :: String -> StringAnnotation
+  StringAnnotation :: !(Maybe SrcLoc) -> String -> StringAnnotation
 
 instance StackAnnotation StringAnnotation where
-  displayStackAnnotation (StringAnnotation str) = str
+  displayStackAnnotationShort (StringAnnotation _srcLoc str) =
+    str
+
+  stackAnnotationSourceLocation (StringAnnotation srcLoc _str) =
+    srcLoc
 
 -- | Use the 'Show' instance of a type to display as the 'StackAnnotation'.
 data ShowAnnotation where
-  ShowAnnotation :: forall a . Show a => a -> ShowAnnotation
+  ShowAnnotation :: forall a . Show a => !(Maybe SrcLoc) -> a -> ShowAnnotation
 
 instance StackAnnotation ShowAnnotation where
-  displayStackAnnotation (ShowAnnotation showAnno) = show showAnno
+  displayStackAnnotationShort (ShowAnnotation _srcLoc showAnno) =
+    show showAnno
+
+  stackAnnotationSourceLocation (ShowAnnotation srcLoc _showAnno) =
+    srcLoc
 
 -- | A 'CallStack' stack annotation.
+--
+-- Captures the whole 'CallStack'.
 newtype CallStackAnnotation = CallStackAnnotation CallStack
 
 instance Show CallStackAnnotation where
@@ -154,9 +144,23 @@ instance Show CallStackAnnotation where
 
 -- | Displays the first entry of the 'CallStack'
 instance StackAnnotation CallStackAnnotation where
-  displayStackAnnotation (CallStackAnnotation cs) = case getCallStack cs of
+  stackAnnotationSourceLocation (CallStackAnnotation cs) =
+    callStackHeadSrcLoc cs
+
+  displayStackAnnotationShort (CallStackAnnotation cs) =
+    callStackHeadFunctionName cs
+
+callStackHeadSrcLoc :: CallStack -> Maybe SrcLoc
+callStackHeadSrcLoc cs =
+  case getCallStack cs of
+    [] -> Nothing
+    (_, srcLoc):_ -> Just srcLoc
+
+callStackHeadFunctionName :: CallStack -> String
+callStackHeadFunctionName cs =
+  case getCallStack cs of
     [] -> "<unknown source location>"
-    ((fnName,srcLoc):_) -> fnName ++ ", called at " ++ prettySrcLoc srcLoc
+    (fnName, _):_ -> fnName
 
 -- ----------------------------------------------------------------------------
 -- Annotate the CallStack with custom data
@@ -172,7 +176,7 @@ instance StackAnnotation CallStackAnnotation where
 --
 -- WARNING: forces the evaluation of @b@ to WHNF.
 {-# NOINLINE annotateStack #-}
-annotateStack :: forall a b. (Typeable a, StackAnnotation a) => a -> b -> b
+annotateStack :: forall a b. (HasCallStack, Typeable a, StackAnnotation a) => a -> b -> b
 annotateStack ann b = unsafePerformIO $
   annotateStackIO ann (evaluate b)
 
@@ -196,9 +200,9 @@ annotateCallStack b = unsafePerformIO $ withFrozenCallStack $
 -- information to stack traces.
 --
 -- WARNING: forces the evaluation of @b@ to WHNF.
-annotateStackString :: forall b . String -> b -> b
+annotateStackString :: forall b . HasCallStack => String -> b -> b
 annotateStackString ann =
-  annotateStack (StringAnnotation ann)
+  annotateStack (StringAnnotation (callStackHeadSrcLoc ?callStack) ann)
 
 -- | @'annotateStackShow' showable b@ annotates the evaluation stack of @b@
 -- with the value @showable@.
@@ -207,37 +211,36 @@ annotateStackString ann =
 -- information to stack traces.
 --
 -- WARNING: forces the evaluation of @b@ to WHNF.
-annotateStackShow :: forall a b . (Typeable a, Show a) => a -> b -> b
+annotateStackShow :: forall a b . (HasCallStack, Typeable a, Show a) => a -> b -> b
 annotateStackShow ann =
-  annotateStack (ShowAnnotation ann)
+  annotateStack (ShowAnnotation (callStackHeadSrcLoc ?callStack) ann)
 
 -- | @'annotateStackIO' showable b@ annotates the evaluation stack of @b@
 -- with the value @showable@.
 --
 -- When decoding the call stack, the annotation frames can be used to add more
 -- information to stack traces.
-annotateStackIO :: forall a b . (Typeable a, StackAnnotation a) => a -> IO b -> IO b
+annotateStackIO :: forall a b . (HasCallStack, Typeable a, StackAnnotation a) => a -> IO b -> IO b
 annotateStackIO ann (IO act) =
   IO $ \s -> annotateStack# (SomeStackAnnotation ann) act s
-{-# NOINLINE annotateStackIO #-}
 
 -- | @'annotateStackStringIO' msg b@ annotates the evaluation stack of @b@
 -- with the value @msg@.
 --
 -- When decoding the call stack, the annotation frames can be used to add more
 -- information to stack traces.
-annotateStackStringIO :: forall b . String -> IO b -> IO b
+annotateStackStringIO :: forall b . HasCallStack => String -> IO b -> IO b
 annotateStackStringIO ann =
-  annotateStackIO (StringAnnotation ann)
+  annotateStackIO (StringAnnotation (callStackHeadSrcLoc ?callStack) ann)
 
 -- | @'annotateStackShowIO' msg b@ annotates the evaluation stack of @b@
 -- with the value @msg@.
 --
 -- When decoding the call stack, the annotation frames can be used to add more
 -- information to stack traces.
-annotateStackShowIO :: forall a b . (Show a) => a -> IO b -> IO b
+annotateStackShowIO :: forall a b . (HasCallStack, Show a) => a -> IO b -> IO b
 annotateStackShowIO ann =
-  annotateStackIO (ShowAnnotation ann)
+  annotateStackIO (ShowAnnotation (callStackHeadSrcLoc ?callStack) ann)
 
 -- | @'annotateCallStackIO' b@ annotates the evaluation stack of @b@ with the
 -- current 'callstack'.